requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.55.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.1.0
